// src/types.d.ts
export declare enum StorageType {
    LOCAL = "local",
    GCS = "gcs",
    S3 = "s3",
    B2 = "b2"
}
export declare type JSON = {
    [id: string]: string | number | boolean | string[] | number[] | boolean[] | {
        [id: string]: JSON;
    };
};
export interface IAdapterConfig {
    type: string;
    slug?: boolean;
    bucketName?: string;
}
// ...
export interface ConfigLocal extends IAdapterConfig {
    directory: string;
    mode?: number | string;
}